Runtime: separate storage host functions and tests into separate mod.
