+++ crates/runtime/src/logic/host_functions.rs
@@ -1,4 +1,6 @@
 
 mod blobs;
+mod storage;
 
 pub use blobs::*;
+pub use storage::*;
